requirements.txt
Pillow==9.2.0
google-cloud-vision==3.1.2
